Fixes action calls
