+++ components/brave_rewards/resources/extension/brave_rewards/background/events/tabEvents.ts
@@ -2,13 +2,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import rewardsPanelActions from '../actions/rewardsPanelActions'
+import { getTabData, onTabData } from '../api/tabs_api'
 
 chrome.tabs.onUpdated.addListener((tabId: number, changeInfo: chrome.tabs.TabChangeInfo, tab: chrome.tabs.Tab) => {
   const activeTabIsLoadingTriggered = Boolean(tab.active && changeInfo.status === 'loading')
-  rewardsPanelActions.onTabRetrieved(tab, activeTabIsLoadingTriggered)
+  onTabData(tab, activeTabIsLoadingTriggered)
 })
 
 chrome.tabs.onActivated.addListener((activeInfo: chrome.tabs.TabActiveInfo) => {
-  rewardsPanelActions.onTabId(activeInfo.tabId)
+  getTabData(activeInfo.tabId)
 })
